requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.2.0

# Deep learning and embeddings
transformers>=4.21.0
sentence-transformers>=2.2.0
tokenizers>=0.13.0

# Recommendation specific
lightgbm>=3.3.0
faiss-cpu>=1.7.4
implicit>=0.6.0

# Web serving
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=1.10.0
redis>=4.5.0

# Data processing
scipy>=1.9.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Experiment tracking
wandb>=0.15.0
mlflow>=2.3.0

# Testing and quality
pytest>=7.2.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
joblib>=1.2.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.20.0